+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.11.1] - 2022-12-26
 ### Fixed
+- Always prefer _Typographic Family_ to _Family Name_ when available.
+  [@CryZe](https://github.com/CryZe)
 - Prevent duplicated family names.
 
 ## [0.11.0] - 2022-12-25
@@ -120,7 +124,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Stretch processing. `ttf-parser` was incorrectly parsing this property.
 
-[Unreleased]: https://github.com/RazrFalcon/fontdb/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/RazrFalcon/fontdb/compare/v0.11.1...HEAD
+[0.11.1]: https://github.com/RazrFalcon/fontdb/compare/v0.11.0...v0.11.1
 [0.11.0]: https://github.com/RazrFalcon/fontdb/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/RazrFalcon/fontdb/compare/v0.9.3...v0.10.0
 [0.9.3]: https://github.com/RazrFalcon/fontdb/compare/v0.9.2...v0.9.3